fix: Resolved restaurant deletion confirmation issue
The delete restaurant function returned an error when trying
to confirm deletion. There was a typo in the html file which
caused a text input to be shown in the confirmation page. Finally,
an improper retrieval of data from the html form detected in
views.py. All of these issues were fixed in this commit.