style: Implemented HTML Templating
Base.html is added to serve as the base template for all the html files.
The base.html file uses bootstrap for the styling so that all other html
files will inherit.

All of the html files were on its own and hard to maintain. Templating is
done now so that they will be easier to maintain.