feat: Added capability to store user data in the database
The database setup file is updated in this commit to include user as a table.
Additionally, all the foreign key relationships are updated so that it observes
data integrity. This new table will store the users' information upon login and
also it will be used for authorization and possibly authentication.