style: Separated regular login and googlelogin code
The regular log in and google login were in a single function and were only
separated by an if statement that checks if the request was post of get. It did
not relly handle the regular log in part. Here in this commit, google login is
separate from regular login so that it will have its own implimentation.
Additionally this will also give room if the site will be adding other options
for OAuth logins.