feat: Added logout capability
Defined a logout function that revokes the access token and destroys all the
login_session variables that were created. Users will be redirected to the login
page after logging out. This is without any flash messages yet. As of now, users
can still see and edit all users' data. Restriction and authorization will be
implemented in the next commits.