feat: Protected html pages
Some of the html pages are now only accessible if the user is logged in. While
some pages are publicly available, they are only for viewing purposes. When a
user that is not logged in tries to manipulate any data, the user is redirected
to the log in page since it is required that the user is logged in before doing
any updates.