feat: Added check after user sign in for database storage
The user is being check if a new user or not after authenticating in the google
sign in. If the user is a new user, then credentials such as email and username
will be added to the database as a User. If not, then the user smoothly signs in
without any database interaction.