fix: Fixed issue on first time login
First time log on via google sign in returned an error saying there was no
record found in the database. Which should be the right case since the user is a
first time user. In this commit, the error has been handled for the function to
return just None instead of throwing an error. This will help the program flow
to the next part which is actually adding the new user into the database.