refactor: Deleted unnecessary functions
Some functions defined in the views.py were not being used at any point in the
program. Those ones that were being used are also deleted in this commit. This
is done so that the views functions are in parallel for easier debugging. Some
functions may come back in the next code refactorings.