feat: Added restriction to buttons per user
Buttons to manipulate other users' data are not invisible to the user if they
are not the owneer of the data. This is to prevent users from accidentally
clicking on a button that will return an error saying they're trying to touch a
data that is not theirs. Passing a user object into a template using flask might
be dangerous and unsafe. This will need some consultation.