chore: Hid the unsecure log in form
The log in form other than the google sign in is not fully developed yet. So
here in this commit, it is being removed so it can not be used as a log in path.
Further development will add hashing functions for the regular signup process
and also input validations.