style: Redesigned login/signup page
All changes done are only for design. Functionalities for the login and signup
forms will be developed in the next commits. This change in the homepage design
will cause UI problems on other pages since the website uses Jinja templating.
Some HTML files are connected. These problems will be resolved in next commits
as well.