style: Renovated homapage design
The restaurant homepage is completely redesigned. Instead of a table listing all
the restaurants, there are cards representing the restaurants with their logos
and perhaps a banner on the upper part of the each card. Now, these cards are a
lot for a single page. Later on, the cards will be limited to a few featured
restaurants.