feat: Retrieve menu by restaurant using AngularJS
AngularJS is added to the stack to provide additional templating and to be able
to do asychronous request to the server. As of now, it is only used to retrieve
the menu of a particular restaurant on request. Meaning, only when the user
clicks "View Menu", that's the only time the request is made. This is to
minimize requests to the server. There is no css design provided yet for the
information on the page.