style: Visible update buttons on user's menu items
Menu items of restaurants owned by the currently logged in user will have
visible buttons to edit or delete the item. The logged in user will not be able
to do any edit or deletion on other users' restaurants' items. The buttons are
currently inactive. But on the next commits, the functionalities will be added.