fix: Lag before menu list is rendered
The previously viewed menu shows for a second and then renders the new menu
being viewed. In some cases, this could be slower and the user might see false
information on the menu list. Because of that, the menu should be blank before
the new menu list is rendered on the screen.