style: Properly positioned forms
Those forms that are in templates were not positioned accordingly with the
banner. The banner is fixed and on top but other elements in some templates
don't recognize that the banner is there and thus place themselves exactly on
the banner. In this commit, proper padding is applied to those templates so that
the would place themselves after the banner,